stybabe/requirements.txt
flask
requests
pillow
tensorflow
keras
flask
tensorflow
sqlite3
